Implement optional func to init Store
